refactor(UI): cleaning up the backend code
