--- src/plugin/ui/src/system/RmlEventSystems.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "Core.hpp"
-
-namespace ES::Plugin::UI::System {
-void UpdateMouseMoveEvent(ES::Engine::Core &core);
-}
