+++ src/plugin/ui/src/system/ShaderSystems.cpp
@@ -6,7 +6,7 @@
 
 #include "OpenGL.pch.hpp"
 
-#include "RmlShaderSystems.hpp"
+#include "ShaderSystems.hpp"
 #include "ShaderManager.hpp"
 
 void ES::Plugin::UI::System::LoadShaderVertColor(ES::Engine::Core &core)
