--- src/plugin/ui/src/system/UpdateButtonState.hpp
@@ -1,13 +1,0 @@
-#pragma once
-
-#include "Engine.hpp"
-
-namespace ES::Plugin::UI::System {
-/**
- * @brief System to update the state of a button
- *
- * @param   core   The core to use
- * @param   e   The entity to update
- */
-void UpdateButtonState(ES::Engine::Core &core);
-} // namespace ES::Plugin::UI::System
