--- src/plugin/ui/src/system/UpdateButtonTexture.hpp
@@ -1,10 +1,0 @@
-#pragma once
-
-#include "Engine.hpp"
-
-namespace ES::Plugin::UI::System {
-/**
- * @brief System to update the texture of a button
- */
-void UpdateButtonTexture(ES::Engine::Core &core);
-} // namespace ES::Plugin::UI::System
