+++ src/plugin/ui/src/system/UpdateUI.hpp
@@ -11,5 +11,13 @@ namespace ES::Plugin::UI::System {
  * @return void
  */
 void Update(ES::Engine::Core &core);
+
+/**
+ * @brief Render the loaded document
+ *
+ * @param core A reference to the core engine, used for the contact listener.
+ *
+ * @return void
+ */
 void Render(ES::Engine::Core &core);
 } // namespace ES::Plugin::UI::System
