--- src/plugin/ui/src/utils/CollisionUtils2D.cpp
@@ -1,7 +1,0 @@
-#include "CollisionUtils2D.hpp"
-
-bool ES::Plugin::UI::Utils::Point2DCollidesRect2D(const ES::Plugin::Math::Rect &rect, const glm::vec2 &point)
-{
-    return point.x >= rect.position.x && point.x <= rect.position.x + rect.size.x && point.y >= rect.position.y &&
-           point.y <= rect.position.y + rect.size.y;
-}
