--- src/plugin/ui/src/utils/CollisionUtils2D.hpp
@@ -1,15 +1,0 @@
-#pragma once
-
-#include "Math.hpp"
-
-namespace ES::Plugin::UI::Utils {
-/**
- * @brief Check if a 2D point is inside a 2D rectangle
- *
- * @param   rect    The rectangle to use
- * @param   point   The point to check
- * @return  true    If the point is inside the rectangle
- * @return  false   If the point is outside the rectangle
- */
-bool Point2DCollidesRect2D(const ES::Plugin::Math::Rect &rect, const glm::vec2 &point);
-} // namespace ES::Plugin::UI::Utils
