+++ src/plugin/ui/src/utils/RenderInterface.cpp
@@ -1,30 +1,4 @@
-#include "Backend.hpp"
-
-static void CheckGlError(const std::string &operation_name)
-{
-    GLenum errCode = glGetError();
-
-    if (errCode != GL_NO_ERROR)
-    {
-        const Rml::Pair<GLenum, std::string> errNames[] = {
-            {GL_INVALID_ENUM,      "GL_INVALID_ENUM"     },
-            {GL_INVALID_VALUE,     "GL_INVALID_VALUE"    },
-            {GL_INVALID_OPERATION, "GL_INVALID_OPERATION"},
-            {GL_OUT_OF_MEMORY,     "GL_OUT_OF_MEMORY"    }
-        };
-        std::string message = "''";
-        for (auto &err : errNames)
-        {
-            if (err.first == errCode)
-            {
-                message = err.second;
-                break;
-            }
-        }
-        ES::Utils::Log::Error(
-            fmt::format("RmlUi: OpenGL error during {}. Error code 0x{:x} {}.", operation_name, errCode, message));
-    }
-}
+#include "RenderInterface.hpp"
 
 ES::Plugin::UI::Utils::RenderInterface::RenderInterface(ES::Engine::Core &core) : _core(core){};
 
@@ -36,16 +10,16 @@ void ES::Plugin::UI::Utils::RenderInterface::UseShaderProgram(const entt::hashed
 
     if (program_id == entt::hashed_string(""))
     {
-        if (activeShaderProgram != entt::hashed_string(""))
-            shaderManager.Get(activeShaderProgram).Disable();
+        if (_activeShaderProgram != entt::hashed_string(""))
+            shaderManager.Get(_activeShaderProgram).Disable();
         return;
     }
 
-    if (activeShaderProgram != program_id)
+    if (_activeShaderProgram != program_id)
     {
-        if (activeShaderProgram != entt::hashed_string(""))
-            shaderManager.Get(activeShaderProgram).Disable();
-        activeShaderProgram = program_id;
+        if (_activeShaderProgram != entt::hashed_string(""))
+            shaderManager.Get(_activeShaderProgram).Disable();
+        _activeShaderProgram = program_id;
     }
     shaderManager.Get(program_id).Use();
 }
@@ -53,10 +27,10 @@ void ES::Plugin::UI::Utils::RenderInterface::UseShaderProgram(const entt::hashed
 void ES::Plugin::UI::Utils::RenderInterface::DisableActiveShaderProgram()
 {
     auto &shaderManager = _core.GetResource<ES::Plugin::OpenGL::Resource::ShaderManager>();
-    if (activeShaderProgram != entt::hashed_string(""))
+    if (_activeShaderProgram != entt::hashed_string(""))
     {
-        shaderManager.Get(activeShaderProgram).Disable();
-        activeShaderProgram = entt::hashed_string("");
+        shaderManager.Get(_activeShaderProgram).Disable();
+        _activeShaderProgram = entt::hashed_string("");
     }
 }
 
@@ -67,110 +41,6 @@ Rml::Rectanglei ES::Plugin::UI::Utils::RenderInterface::VerticallyFlipped(Rml::R
     flipped_rect.p0.y = viewport_height - rect.p1.y;
     flipped_rect.p1.y = viewport_height - rect.p0.y;
     return flipped_rect;
-}
-
-bool ES::Plugin::UI::Utils::RenderInterface::CreateFramebuffer(FramebufferData &out_fb, int width, int height,
-                                                               int samples, FramebufferAttachment attachment,
-                                                               GLuint shared_depth_stencil_buffer)
-{
-#ifdef RMLUI_PLATFORM_EMSCRIPTEN
-    constexpr GLint wrap_mode = GL_CLAMP_TO_EDGE;
-#else
-    constexpr GLint wrap_mode = GL_CLAMP_TO_BORDER; // GL_REPEAT GL_MIRRORED_REPEAT GL_CLAMP_TO_EDGE
-#endif
-
-    constexpr GLenum color_format = GL_RGBA8;   // GL_RGBA8 GL_SRGB8_ALPHA8 GL_RGBA16F
-    constexpr GLint min_mag_filter = GL_LINEAR; // GL_NEAREST
-    const Rml::Colourf border_color(0.f, 0.f);
-
-    GLuint framebuffer = 0;
-    glGenFramebuffers(1, &framebuffer);
-    glBindFramebuffer(GL_FRAMEBUFFER, framebuffer);
-
-    GLuint color_tex_buffer = 0;
-    GLuint color_render_buffer = 0;
-    if (samples > 0)
-    {
-        glGenRenderbuffers(1, &color_render_buffer);
-        glBindRenderbuffer(GL_RENDERBUFFER, color_render_buffer);
-        glRenderbufferStorageMultisample(GL_RENDERBUFFER, samples, color_format, width, height);
-        glFramebufferRenderbuffer(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_RENDERBUFFER, color_render_buffer);
-    }
-    else
-    {
-        glGenTextures(1, &color_tex_buffer);
-        glBindTexture(GL_TEXTURE_2D, color_tex_buffer);
-        glTexImage2D(GL_TEXTURE_2D, 0, color_format, width, height, 0, GL_RGBA, GL_UNSIGNED_BYTE, NULL);
-
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, min_mag_filter);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, min_mag_filter);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, wrap_mode);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, wrap_mode);
-#ifndef RMLUI_PLATFORM_EMSCRIPTEN
-        glTexParameterfv(GL_TEXTURE_2D, GL_TEXTURE_BORDER_COLOR, &border_color[0]);
-#endif
-
-        glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, color_tex_buffer, 0);
-    }
-
-    // Create depth/stencil buffer storage attachment.
-    GLuint depth_stencil_buffer = 0;
-    if (attachment != FramebufferAttachment::None)
-    {
-        if (shared_depth_stencil_buffer)
-        {
-            // Share depth/stencil buffer
-            depth_stencil_buffer = shared_depth_stencil_buffer;
-        }
-        else
-        {
-            // Create new depth/stencil buffer
-            glGenRenderbuffers(1, &depth_stencil_buffer);
-            glBindRenderbuffer(GL_RENDERBUFFER, depth_stencil_buffer);
-
-            glRenderbufferStorageMultisample(GL_RENDERBUFFER, samples, GL_DEPTH24_STENCIL8, width, height);
-        }
-
-        glFramebufferRenderbuffer(GL_FRAMEBUFFER, GL_DEPTH_STENCIL_ATTACHMENT, GL_RENDERBUFFER, depth_stencil_buffer);
-    }
-
-    const GLuint framebuffer_status = glCheckFramebufferStatus(GL_FRAMEBUFFER);
-    if (framebuffer_status != GL_FRAMEBUFFER_COMPLETE)
-    {
-        ES::Utils::Log::Error(
-            fmt::format("OpenGL framebuffer could not be generated. Error code {:x}.", framebuffer_status));
-        return false;
-    }
-
-    glBindFramebuffer(GL_FRAMEBUFFER, 0);
-    glBindTexture(GL_TEXTURE_2D, 0);
-    glBindRenderbuffer(GL_RENDERBUFFER, 0);
-
-    out_fb = {};
-    out_fb.width = width;
-    out_fb.height = height;
-    out_fb.framebuffer = framebuffer;
-    out_fb.color_tex_buffer = color_tex_buffer;
-    out_fb.color_render_buffer = color_render_buffer;
-    out_fb.depth_stencil_buffer = depth_stencil_buffer;
-    out_fb.owns_depth_stencil_buffer = !shared_depth_stencil_buffer;
-
-    // CheckGlError("CreateFramebuffer");
-    return true;
-}
-
-void ES::Plugin::UI::Utils::RenderInterface::DestroyFramebuffer(FramebufferData &fb)
-{
-    if (fb.framebuffer)
-        glDeleteFramebuffers(1, &fb.framebuffer);
-    if (fb.color_tex_buffer)
-        glDeleteTextures(1, &fb.color_tex_buffer);
-    if (fb.color_render_buffer)
-        glDeleteRenderbuffers(1, &fb.color_render_buffer);
-    if (fb.owns_depth_stencil_buffer && fb.depth_stencil_buffer)
-        glDeleteRenderbuffers(1, &fb.depth_stencil_buffer);
-    fb = {};
-    // CheckGlError("DestroyFrameBuffer");
 }
 
 Rml::CompiledGeometryHandle
@@ -210,8 +80,6 @@ ES::Plugin::UI::Utils::RenderInterface::CompileGeometry(Rml::Span<const Rml::Ver
 
     glBindVertexArray(0);
     glBindBuffer(GL_ARRAY_BUFFER, 0);
-
-    // CheckGlError("CompileGeometry");
 
     std::string key = "rml_mesh_" + std::to_string(_next_geom_id);
     CompiledGeometryData geometry;
@@ -276,8 +144,6 @@ void ES::Plugin::UI::Utils::RenderInterface::RenderGeometry(Rml::CompiledGeometr
 
     glBindVertexArray(0);
     glBindTexture(GL_TEXTURE_2D, 0);
-
-    // CheckGlError("RenderCompileGeometry");
 }
 
 void ES::Plugin::UI::Utils::RenderInterface::SetTransform(const Rml::Matrix4f *new_transform)
@@ -387,8 +253,6 @@ Rml::TextureHandle ES::Plugin::UI::Utils::RenderInterface::CreateTexture(Rml::Sp
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_REPEAT);
     glBindTexture(GL_TEXTURE_2D, 0);
 
-    // CheckGlError("CreateTexture");
-
     return static_cast<Rml::TextureHandle>(texture_id);
 }
 
@@ -436,7 +300,6 @@ void ES::Plugin::UI::Utils::RenderInterface::SetScissor(Rml::Rectanglei region, 
     }
 
     _scissor_state = region;
-    // CheckGlError("SetSissor");
 }
 
 void ES::Plugin::UI::Utils::RenderInterface::BeginFrame()
@@ -522,13 +385,12 @@ void ES::Plugin::UI::Utils::RenderInterface::BeginFrame()
     _scissor_state = Rml::Rectanglei::MakeInvalid();
 
     glBindFramebuffer(GL_FRAMEBUFFER, 0);
-    // CheckGlError("BeginFrame");
 }
 
 void ES::Plugin::UI::Utils::RenderInterface::EndFrame()
 {
-    const FramebufferData &fb_active = _render_layers.GetTopLayer();
-    const FramebufferData &fb_postprocess = _render_layers.GetPostprocessPrimary();
+    const RenderLayerStack::FramebufferData &fb_active = _render_layers.GetTopLayer();
+    const RenderLayerStack::FramebufferData &fb_postprocess = _render_layers.GetPostprocessPrimary();
     const auto &windowSize = _core.GetResource<ES::Plugin::Window::Resource::Window>().GetSize();
 
     // Resolve MSAA to postprocess framebuffer.
@@ -604,6 +466,4 @@ void ES::Plugin::UI::Utils::RenderInterface::EndFrame()
     glStencilMaskSeparate(GL_BACK, _glstate_backup.stencil_back.writemask);
     glStencilOpSeparate(GL_BACK, _glstate_backup.stencil_back.fail, _glstate_backup.stencil_back.pass_depth_fail,
                         _glstate_backup.stencil_back.pass_depth_pass);
-
-    // CheckGlError("EndFrame");
 }
